fixed warning

+++ engine/tinygl/zline.cpp
@@ -8,13 +8,13 @@ void ZB_plot(ZBuffer * zb, ZBufferPoint * p) {
 	unsigned short *pz;
 	unsigned long *pz_2;
 	PIXEL *pp;
-	int zz;
+	unsigned int zz;
 
 	pz = zb->zbuf + (p->y * zb->xsize + p->x);
 	pz_2 = zb->zbuf2 + (p->y * zb->xsize + p->x);
 	pp = (PIXEL *)((char *) zb->pbuf + zb->linesize * p->y + p->x * PSZB);
 	zz = p->z >> ZB_POINT_Z_FRAC_BITS;
-	if ((ZCMP(zz, *pz)) && (ZCMP(p->z, *pz_2)) ) {
+	if ((ZCMP(zz, *pz)) && (ZCMP((unsigned int)p->z, *pz_2)) ) {
 		*pp = RGB_TO_PIXEL(p->r, p->g, p->b);
 		*pz_2 = p->z;
     }
